Добавляет модули и ф-цию из ДЗ 5.16
